chore(architecture): document Layer 3 module archetypes and close structural gaps
Resolves audit-architecture finding that Layer 3 submodules followed two
incompatible conventions without documentation. Chose the "document" path
(issue option 2) over forcing the triple onto stateful auth/runtime
services, which would be harmful.

- Add "Layer 3 Module Archetypes" section to docs/architecture/layer-architecture.md
  documenting three patterns (pure-function triple, stateful service, thin
  utility) plus a hybrid-in-progress carve-out and a quick-chooser table.
- Cross-reference the archetype section from CLAUDE.md file-organization.
- Add crates/octarine/src/security/queries/types.rs using the 3-line
  `pub use` pattern from security/formats/types.rs — every security/*
  submodule now has a local types.rs.
- Split crates/octarine/src/primitives/identifiers/types.rs (469 lines)
  into a types/ directory (core.rs, personal.rs, financial.rs) mirroring
  the Layer 3 identifiers/types/ layout.

Documentation + structural refactor only — no API changes, no behavior
changes. Full preflight (fmt + clippy + arch-check + 6380 unit tests)
passes.

Closes #175
